+++ lektor/project.py
@@ -5,6 +5,7 @@ import sys
 from inifile import IniFile
 from werkzeug.utils import cached_property
 
+from lektor.environment import Environment
 from lektor.utils import comma_delimited
 from lektor.utils import get_cache_dir
 from lektor.utils import untrusted_to_os_path
@@ -139,8 +140,6 @@ class Project:
 
     def make_env(self, load_plugins=True):
         """Create a new environment for this project."""
-        from lektor.environment import Environment
-
         return Environment(self, load_plugins=load_plugins)
 
     @cached_property
